+++ src/gmp.c
@@ -268,48 +268,6 @@ check_certificate (const char *cert_str, const char *credential_type)
 }
 
 /**
- * @brief Check that a string represents a valid Private Key.
- *
- * @param[in]  key_str      Private Key string.
- * @param[in]  key_phrase   Private Key passphrase.
- *
- * @return 0 if valid, 1 otherwise.
- */
-static int
-check_private_key (const char *key_str, const char *key_phrase)
-{
-  gnutls_x509_privkey_t key;
-  gnutls_datum_t data;
-  int ret;
-
-  assert (key_str);
-  if (gnutls_x509_privkey_init (&key))
-    return 1;
-  data.size = strlen (key_str);
-  data.data = (void *) g_strdup (key_str);
-  ret = gnutls_x509_privkey_import2 (key, &data, GNUTLS_X509_FMT_PEM,
-                                     key_phrase, 0);
-  if (ret)
-    {
-      gchar *public_key;
-      public_key = gvm_ssh_public_from_private (key_str, key_phrase);
-
-      if (public_key == NULL)
-        {
-          gnutls_x509_privkey_deinit (key);
-          g_free (data.data);
-          g_message ("%s: import failed: %s",
-                     __func__, gnutls_strerror (ret));
-          return 1;
-        }
-      g_free (public_key);
-    }
-  g_free (data.data);
-  gnutls_x509_privkey_deinit (key);
-  return 0;
-}
-
-/**
  * @brief Check that a string represents a valid Public Key.
  *
  * @param[in]  key_str  Public Key string.